refactor(components): Rename InputVariables to ControlVariables
