+++ apps/web/src/pages/templates/editor_v2/TemplateStepEditorV2.tsx
@@ -3,7 +3,7 @@ import { useMutation, useQuery } from '@tanstack/react-query';
 import { useEffect, useState } from 'react';
 import { WorkflowsPageTemplate, WorkflowsPanelLayout } from '../../../studio/components/workflows/layout';
 import { WorkflowStepEditorContentPanel } from '../../../studio/components/workflows/step-editor/WorkflowStepEditorContentPanel';
-import { WorkflowStepEditorInputsPanel } from '../../../studio/components/workflows/step-editor/WorkflowStepEditorInputsPanel';
+import { WorkflowStepEditorControlsPanel } from '../../../studio/components/workflows/step-editor/WorkflowStepEditorControlsPanel';
 import { useTemplateController } from '../components/useTemplateController';
 import { api } from '../../../api';
 import { WORKFLOW_NODE_STEP_ICON_DICTIONARY } from '../../../studio/components/workflows/node-view/WorkflowNodes';
@@ -67,7 +67,7 @@ export const WorkflowsStepEditorPageV2 = () => {
     <WorkflowsPageTemplate title={title} icon={<Icon size="32" />}>
       <WorkflowsPanelLayout>
         <WorkflowStepEditorContentPanel error={error} step={step} preview={preview} isLoadingPreview={loadingPreview} />
-        <WorkflowStepEditorInputsPanel
+        <WorkflowStepEditorControlsPanel
           isLoadingSave={isSavingControls}
           onSave={() => {
             onControlsSave();
